Refine drag scroll
Scroll seems to be misbehaving when I horizontally scroll before vertical scroll, and when using the mouse as scroll control, it's tricky to scroll only vertically. Therefore, separating the "multi-axis scroll" from the vertical scroll, since that seems to work better to more consistently vertically scroll.

Also, having to hold down a key while scrolling is a bit annoying, so switching to a toggle setup.