Horizontal joystick scrolling on layer 1
From previous usages, I found that enabling horizontal scrolling on certan apps/pages causes issues, and with F360 I don't want to use h-scroll, so by default it's locked to only vertical scrolling.

However, it'd still be useful to have in apps that do handle h-scrolling well (e.g. figma), so I'm adding it, but only when layer 1 is active.